Fixes #15, it's an issue in bootstrap-wyishtml5.
Enter/Return key produces <br> tag rather than <p> tag.
https://github.com/jhollingworth/bootstrap-wysihtml5/issues/159

Django-wysihtml5 provides a setting called WYSIHTML5_ALLOWED_TAGS to filter
the HTML tags that are allowed to hit the content of the Django field.
It's a security measurement to avoid code injection and none valid tags.

It wasn't explained in the README.rst and not in use in the demo site. This
commit adds the <br> as a valid tag for the demo site (it has to be fixed
in bootstrap-wysihtml5) and updates the README.rst file.